Read config parameters from file
The configuration parameters where hard coded at several places in different files.
Although the export process and its parameters were well documented in Readme.md it was not
that easy to find all the places in code that needed adaption. Furthermore configuration changes
needed a recompilation of the app.

Most configuration parameters have been concentrated into a new struct in Config.swift by this commit.

Additionally the process will try to read the config from a json file in the user's Application Support folder.
If such file does not exist the process will create a default config file and print some information how to
setup the user's config file.

The logic of the export process itself is not changed by this commit.